==> 7_b_information.ipynb <==
website = "https://vimergy.com/"

budget = 40000
aov=80
ctr=1.5
target_gender = "90% female"
target_hhi = "+$100k"


ai_backend_brand_base_url = "http://localhost:3001/api/v1/brand"
headers = {
    "authorization": "testtoken0001"
}

import requests
# get data from apollo


def get_brand_info(website):
    api_url = f"{ai_backend_brand_base_url}/info"
    try:
        response = requests.post(api_url, json={"website_url": website}, headers=headers)
        brand_info = response.json()
        return brand_info
    except Exception as e:
        print(f"Error getting brand info for {website}: {e}")
        return None








import json
brand_info = get_brand_info(website)


with open("branding_info.json", "w") as f:
    json.dump(brand_info, f)
